Move files to other packages
