+++ synctools/src/androidTest/kotlin/at/bitfire/synctools/storage/jtx/JtxBatchOperationTest.kt
@@ -10,10 +10,10 @@ import androidx.core.content.contentValuesOf
 import androidx.test.platform.app.InstrumentationRegistry
 import at.bitfire.synctools.storage.BatchOperation
 import at.bitfire.synctools.storage.TaskProvider
-import at.bitfire.synctools.storage.asSyncAdapter
 import at.bitfire.synctools.test.BuildConfig
 import at.bitfire.synctools.test.GrantPermissionOrSkipRule
 import at.techbee.jtx.JtxContract
+import at.techbee.jtx.JtxContract.asSyncAdapter
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
