+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/calendar/builder/AndroidRecurrenceMapper.kt
@@ -4,9 +4,9 @@
 
 package at.bitfire.synctools.mapping.calendar.builder
 
-import at.bitfire.synctools.icalendar.DateUtils
 import at.bitfire.synctools.util.AndroidTimeUtils.toInstant
 import at.bitfire.synctools.util.AndroidTimeUtils.toZonedDateTime
+import at.bitfire.synctools.util.DateUtils
 import at.bitfire.synctools.util.TimeApiExtensions.toLocalDate
 import net.fortuna.ical4j.model.Property
 import java.time.LocalDate
