+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/calendar/AndroidCalendarProvider.kt
@@ -17,7 +17,6 @@ import androidx.annotation.VisibleForTesting
 import androidx.core.content.contentValuesOf
 import at.bitfire.synctools.icalendar.Css3Color
 import at.bitfire.synctools.storage.LocalStorageException
-import at.bitfire.synctools.storage.asSyncAdapter
 import at.bitfire.synctools.storage.calendar.AndroidCalendarProvider.Companion.COLUMN_CALENDAR_SYNC_STATE
 import at.bitfire.synctools.storage.toContentValues
 import java.util.LinkedList
