+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/tasks/DmfsTaskListProvider.kt
@@ -11,7 +11,6 @@ import android.content.ContentValues
 import android.os.RemoteException
 import at.bitfire.synctools.storage.LocalStorageException
 import at.bitfire.synctools.storage.TaskProvider
-import at.bitfire.synctools.storage.asSyncAdapter
 import at.bitfire.synctools.storage.toContentValues
 import org.dmfs.tasks.contract.TaskContract
 import java.util.LinkedList
